Moves aci functions to a separate hpp file
